Move namespaces around to better match purpose
