+++ osu.Server.Queues.ScorePump/Performance/UpdateUserTotalsCommands.cs
@@ -2,15 +2,14 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using McMaster.Extensions.CommandLineUtils;
-using osu.Server.Queues.ScorePump.Performance.Totals;
-using osu.Server.Queues.ScorePump.Performance.Values;
+using osu.Server.Queues.ScorePump.Performance.UserTotals;
 
 namespace osu.Server.Queues.ScorePump.Performance
 {
-    [Command(Name = "update-pp", Description = "Computes the performance (pp) of scores.")]
-    [Subcommand(typeof(UpdateScoresCommands))]
-    [Subcommand(typeof(UpdateUserTotalsCommands))]
-    public sealed class PerformanceCommands
+    [Command("totals", Description = "Updates user total PP values.")]
+    [Subcommand(typeof(UpdateAllUserTotals))]
+    [Subcommand(typeof(UpdateUserTotalsForUsers))]
+    public sealed class UpdateUserTotalsCommands
     {
         public int OnExecute(CommandLineApplication app)
         {
