+++ osu.Server.Queues.ScorePump/Performance/UserTotals/UpdateAllUserTotals.cs
@@ -9,7 +9,7 @@ using Dapper;
 using McMaster.Extensions.CommandLineUtils;
 using osu.Server.Queues.ScoreStatisticsProcessor;
 
-namespace osu.Server.Queues.ScorePump.Performance.Totals
+namespace osu.Server.Queues.ScorePump.Performance.UserTotals
 {
     [Command("all", Description = "Updates the total PP of all users.")]
     public class UpdateAllUserTotals : PerformanceCommand
