+++ osu.Server.Queues.ScorePump/Program.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
-using osu.Server.Queues.ScorePump.Performance;
+using osu.Server.Queues.ScorePump.Queue;
 
 namespace osu.Server.Queues.ScorePump
 {
