+++ osu.Server.Queues.ScorePump/Queue/ClearQueue.cs
@@ -5,7 +5,7 @@ using System;
 using System.Threading;
 using McMaster.Extensions.CommandLineUtils;
 
-namespace osu.Server.Queues.ScorePump
+namespace osu.Server.Queues.ScorePump.Queue
 {
     [Command("clear-queue", Description = "Completely empties the processing queue")]
     public class ClearQueue : QueueCommand
