+++ osu.Server.Queues.ScorePump/Queue/PumpAllScores.cs
@@ -7,7 +7,7 @@ using Dapper;
 using McMaster.Extensions.CommandLineUtils;
 using osu.Server.Queues.ScoreStatisticsProcessor.Models;
 
-namespace osu.Server.Queues.ScorePump
+namespace osu.Server.Queues.ScorePump.Queue
 {
     [Command("all", Description = "Pumps scores through the queue for reprocessing")]
     public class PumpAllScores : QueueCommand
