+++ osu.Server.Queues.ScorePump/Queue/PumpTestData.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using McMaster.Extensions.CommandLineUtils;
 using osu.Server.Queues.ScoreStatisticsProcessor.Models;
 
-namespace osu.Server.Queues.ScorePump
+namespace osu.Server.Queues.ScorePump.Queue
 {
     [Command("test", Description = "Pumps empty test scores to the queue")]
     public class PumpTestData : QueueCommand
